pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Rem M. <morrig407@gmail.com>"]
readme = "README.md"
packages = [{include = "page_analyzer"}]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^3.0.3"
gunicorn = "^22.0.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
validators = "^0.29.0"
requests = "^2.32.3"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.0"
pytest = "^8.2.2"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
